requirements.txt
colorthief
cryptg
Pillow
Telethon
tqdm
